--- src/components/Biospecimens/Request/requestBiospecimen.module.scss
@@ -1,27 +1,0 @@
-@import 'src/style/themes/include/colors';
-
-.modalWrapper {
-  .description {
-    margin-bottom: 16px;
-  }
-
-  .nameForm {
-    .note {
-      font-size: 12px;
-      color: $gray-8;
-      margin-bottom: 8px;
-    }
-  }
-}
-
-.notStriped {
-  background-color: inherit !important;
-}
-
-.tooltip {
-  border-bottom: 1px dotted;
-}
-
-.alertTitle {
-  font-weight: 600;
-}
