--- src/views/Dashboard/components/DashboardCards/BiospecimenRequests/editBiospecimenRequestModal.module.scss
@@ -1,3 +1,0 @@
-.formItem {
-  margin-bottom: 0;
-}
