--- src/views/Dashboard/components/DashboardCards/BiospecimenRequests/index.module.scss
@@ -1,21 +1,0 @@
-@import 'src/style/themes/include/colors';
-
-.biospecimenRequestWrapper {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-
-  .biospecimenRequestList {
-    height: 100%;
-    overflow: auto;
-    min-height: 200px;
-
-    div[class$='-item-meta'] {
-      margin-bottom: 0;
-    }
-
-    *[class$='-item-meta-description'] {
-      font-size: 12px;
-    }
-  }
-}
