--- src/views/Dashboard/components/DashboardCards/BiospecimenRequests/listItem.module.scss
@@ -1,4 +1,0 @@
-.title > span::after {
-  content: '';
-  display: block;
-}
